requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.3
croniter==0.3.30
cryptography==2.4.2
dj-database-url==0.5.0
Django==2.1.4
django-environ==0.4.5
django-extensions==2.1.4
django-redis==4.10.0
django-redis-cache==2.0.0
django-rq==2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
jwt==0.5.4
more-itertools==4.3.0
Pillow==5.3.0
pluggy==0.8.0
postmarker==0.13.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
PyPDF2==1.26.0
pytest==4.0.2
pytest-cov==2.7.1
pytest-django==3.4.4
python-coveralls==2.9.1
python-dateutil==2.8.0
python-redis-lock==3.3.1
pytz==2018.7
PyYAML==5.1
redis==3.2.1
redis-cache==0.1.5
redis-structures==0.1.7
requests==2.21.0
rq==1.0
rq-scheduler==0.9
six==1.12.0
urllib3==1.24.3
vital-tools==0.1.13
whitenoise==4.1.2
